// UI/AngularDemo.Web/src/app/models/student.ts
export interface Student {
  id?: number;
  schoolId?: number;
  schoolName?: string;
  schoolCode?: string;
  rollNumber: string;
  fullName: string;
  userName?: string;
  password?: string;
  dateOfBirth: string;
  gender?: string;
  classMasterId?: number;
  sectionId?: number;
  academicYearId?: number;
  admissionDate: string;
  email?: string;
  phone?: string;
  address?: string;
  isActive: boolean;
  createdDate?: string;
}